requirements.txt
# TrendPilot AI - Python 3.10+
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0
openai>=1.12.0
chromadb>=0.4.22
python-dotenv>=1.0.0
